+++ Sources/Toki/main.swift
@@ -1,5 +1,9 @@
 import AppKit
 
+if UpdateInstaller.runHelperIfRequested(arguments: CommandLine.arguments) {
+    exit(EXIT_SUCCESS)
+}
+
 let app = NSApplication.shared
 let delegate = AppDelegate()
 app.delegate = delegate